refactor(tls): move client certificate-store helpers off the config types
Review feedback: configuration types should not carry functionality.
TlsClient::create_certificate_store and its helpers move to
services/subgraph/http, next to generate_tls_client_config which already
serves the same three consumers (subgraph/connector HTTP clients and the
Redis client).
